StartScreensaver
Check if the mouse/keyboard is idle and than run screensaver start command.
If a YouTube video trough Safari is playing, the screensaver does not start. 

Useful if you use programs such as eqMac for driving sound to a virtual interface: this type of program keep the computer always active and the screensaver can't start automatically.